Reorganização do apiservice associado a admin + remoção código desnecessário
